+++ ext/stub.h
@@ -111,6 +111,7 @@
 #include "stub/issue2324extended.zep.h"
 #include "stub/issue2385.zep.h"
 #include "stub/issue2394.zep.h"
+#include "stub/issue2428.zep.h"
 #include "stub/issue2469.zep.h"
 #include "stub/issue2505crosschain.zep.h"
 #include "stub/issue2505extended.zep.h"